requirements.txt
python3-indy>=1.16.0
python-dotenv>=0.19.0
asyncio>=3.4.3
aiohttp>=3.8.1
